Update README.md
The idea for the app is to allow users to set up short/ long term challenges and be able to sustain their motivation. We recognize that people usually set up goals out of a whim, and fail to accomplish them because this temporary motivation wears out over time. Our app captures the user?s impulse in form of images and texts. When the user is motivated, they set up a challenge with the application and attach with it a list of inspirations that they can look back at in the future. When they feel unmotivated to continue to accomplish this goal, they can revisit the message that their previous selves have left them.
The application is entirely built in objective C, with Core Data as the backend. Storyboard and Autolayout is used for front end.